Refactor to add ability for Register size specification in type signature
RegisterName is now a struct which is inherited by RegisterName8 and
RegisterName16 for 8 and 16 bit registers. All register constants are
now static constexpr static members of the RegisterNames struct. All
previous register constants like RegisterName::A are now under the
RegisterNames scope, (RegisterNames::A).

This change was made to allow for type level distinctions between 8 bit
and 16 bit registers. For example, a LoadEvent16 operation can now
specifically ask for a Register16 as the destination instead of just any
register.

This change/refactor brought to light some interesting problems. C++ as
of this commit does not have the concept of enum supersets/subsets so
this solution was used instead.